% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unet.R
\name{unet_decoder_block}
\alias{unet_decoder_block}
\title{U-Net Decoder Block}
\usage{
unet_decoder_block(in_channels, out_channels, use_dropout = FALSE)
}
\arguments{
\item{in_channels}{Number of input channels (after skip-concat).}

\item{out_channels}{Number of output channels.}

\item{use_dropout}{Whether to apply 50% dropout (used in first decoder blocks).}
}
\value{
An `nn_module` instance.
}
\description{
U-Net Decoder Block
}
