% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unet.R
\name{unet_encoder_block}
\alias{unet_encoder_block}
\title{U-Net Encoder Block}
\usage{
unet_encoder_block(in_channels, out_channels, use_batchnorm = TRUE)
}
\arguments{
\item{in_channels}{Number of input channels.}

\item{out_channels}{Number of output channels.}

\item{use_batchnorm}{Whether to apply BatchNorm (default TRUE).}
}
\value{
An `nn_module` instance.
}
\description{
A single downsampling block: Conv2d -> BatchNorm -> LeakyReLU.
The first encoder block skips BatchNorm (standard U-Net convention).
}
