% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unet.R
\name{unet_model}
\alias{unet_model}
\title{U-Net Generator}
\usage{
unet_model(in_channels = 6L, out_channels = 6L, filters = 64L)
}
\arguments{
\item{in_channels}{Number of input channels. Default `6` (Landsat-8 bands).}

\item{out_channels}{Number of output channels. Default `6` (Sentinel-2 bands).}

\item{filters}{Base number of filters. Default `64`.}
}
\value{
An `nn_module` instance.
}
\description{
A U-Net with 4 encoder blocks, a bottleneck, and 4 decoder blocks.
Landsat-8 imagery (default 6 bands) to Sentinel-2 (default 6 bands).
}
\details{
Architecture (default filters = 64):
}
